requirements.txt
flask
python-slugify